Revert "Move over_segment query to new ::Query module and namespace"
This reverts commit b1ac098dcaa3919546544897564066948d76dde4.
