--- app/models/query/base.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Query
-  class Base
-    def self.sql_for_existing_scope(scope)
-      scope.connection.unprepared_statement { scope.reorder(nil).select('id').to_sql }
-    end
-  end
-end
